requirements.txt
-e git+ssh://git@gitlab.fel.cvut.cz/factchecking/experimental-vaclav-hlavac.git@c03e16c0533029619d61f76cf079d044c4a761ef#egg=afc
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
Babel==2.10.1
backcall==0.2.0
beautifulsoup4==4.11.1
black==22.6.0
bleach==5.0.1
blis==0.7.8
catalogue==2.0.8
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.2
coloredlogs==15.0.1
commonmark==0.9.1
cycler==0.11.0
cymem==2.0.6
Cython==0.29.32
datasets==2.1.0
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.4
docker-pycreds==0.4.0
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.4.0/en_core_web_lg-3.4.0-py3-none-any.whl
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.4.0/en_core_web_md-3.4.0-py3-none-any.whl
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.0/en_core_web_sm-3.4.0-py3-none-any.whl
entrypoints==0.4
executing==0.8.3
faiss-cpu==1.7.2
fastjsonschema==2.15.3
filelock==3.6.0
flatbuffers==2.0
fonttools==4.33.2
frozenlist==1.3.0
fsspec==2022.5.0
gitdb==4.0.9
GitPython==3.1.27
graphviz==0.20.1
huggingface-hub==0.5.1
humanfriendly==10.0
idna==3.3
importlib-metadata==4.12.0
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
json5==0.9.6
jsonschema==4.7.2
jupyter-client==7.3.4
jupyter-core==4.11.1
jupyter-server==1.16.0
jupyterlab==3.3.4
jupyterlab-pygments==0.2.2
jupyterlab-server==2.13.0
jupyterlab-vim==0.15.1
jupyterlab-widgets==1.1.1
kiwisolver==1.4.3
langcodes==3.3.0
lightgbm==3.3.2
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
multidict==6.0.2
multiprocess==0.70.12.2
murmurhash==1.0.7
mypy-extensions==0.4.3
nbclassic==0.3.7
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
nltk==3.7
nmslib==2.1.1
notebook==6.4.12
notebook-shim==0.1.0
numpy==1.22.4
onnxruntime==1.12.1
packaging==21.3
pandarallel==1.6.3
pandas==1.5.1
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pathtools==0.1.2
pathy==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.2
preshed==3.0.6
prometheus-client==0.14.1
promise==2.3
prompt-toolkit==3.0.30
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==7.0.0
pybind11==2.6.1
pycparser==2.21
pydantic==1.9.1
Pygments==2.12.0
pyjnius==1.4.2
pyparsing==3.0.9
pyrsistent==0.18.1
pyserini==0.17.0
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
pyzmq==23.2.0
regex==2022.3.15
requests==2.27.1
responses==0.18.0
rich==12.4.4
sacremoses==0.0.49
seaborn==0.11.2
Send2Trash==1.8.0
sentence-transformers==2.2.2
sentencepiece==0.1.97
sentry-sdk==1.5.10
setproctitle==1.2.3
shortuuid==1.0.8
six==1.16.0
smart-open==5.2.1
smmap==5.0.0
sniffio==1.2.0
soupsieve==2.3.2.post1
spacy==3.4.1
spacy-legacy==3.0.9
spacy-loggers==1.0.3
srsly==2.4.4
stack-data==0.3.0
sympy==1.10.1
terminado==0.15.0
thinc==8.1.0
threadpoolctl==3.1.0
tinycss2==1.1.1
tokenizers==0.12.1
tomli==2.0.1
torchviz==0.0.2
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
transformers==4.18.0
typer==0.4.2
typing_extensions==4.4.0
urllib3==1.26.9
wandb==0.12.20
wasabi==0.10.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.2
widgetsnbextension==3.6.1
xxhash==3.0.0
yarl==1.7.2
zipp==3.8.0
